bugfix mash pipeline step5

+++ scripts/createMashPlans.sh
@@ -285,8 +285,11 @@ then
         
         for x in pacbio/*.fasta.gz 10x/*.fastq.gz hic/*.fastq.gz
         do
-        	out=screen/${x%.fast[aq].gz}
-        	echo "mash screen -p ${threads} -w ${MASH_REF_GENOMES}/refseq.genomes.k21s1000.msh ${x} > ${out}"
+        	if [[ -f ${x} ]]
+        	then
+        		out=screen/$(basename ${x%.fast[aq].gz})
+        		echo "mash screen -p ${threads} -w ${MASH_REF_GENOMES}/refseq.genomes.k21s1000.msh ${x} > ${out}"
+        	fi
 		done > mash_05_mashScreen_block_${RAW_DB%.db}.${slurmID}.plan   
 		echo "mash $(${PACBIO_BASE_ENV} && mash --version && ${PACBIO_BASE_ENV_DEACT})" > mash_05_mashScreen_block_${RAW_DB%.db}.${slurmID}.version     
     else
